requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.35.2
torch==2.1.1
python-multipart==0.0.6
arxiv==2.0.0
grobid-client-python==0.0.5
pdfminer.six==20221105
numpy==1.24.3
python-dotenv==1.0.0
pandas==2.1.0
tqdm==4.66.1
scikit-learn==1.3.0
python-magic==0.4.27
nltk==3.8.1 